moved all img to public folder to solve error in console
