+++ public/git-merge.png
