+++ src/data/projects.json
@@ -28,7 +28,7 @@
     "live": "https://library-recipe.netlify.app/",
     "code": "https://github.com/JeffieJansson/recipe-library",
     "image": {
-      "src": "src/assets/recipe.jpg",
+      "src": "/recipe.jpg",
       "alt": "pictures of food recipes and filtering options"
     }
   },
@@ -45,7 +45,7 @@
     "live": "https://project-weather-app-b2.netlify.app/",
     "code": "https://github.com/marinalendt-png/js-project-weather-app",
     "image": {
-      "src": "src/assets/weather.jpg",
+      "src": "/weather.jpg",
       "alt": "picture of weather app"
     }
   },
